fix problem

+++ src/main/java/The/Geeks/RESM/repositories/RoleRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import The.Geeks.RESM.domain.Role;
 
-public interface RoleRepo extends JpaRepository <Role,Long> {
+public interface RoleRepo extends JpaRepository <Role,Integer> {
     Role  findByName(String name);
 
 }
